fix(map,selection): selection now works when the map is scrolled
This is a tricky one - the map has the outer container which has the scrollbar and the inner container which overflows.
Event handlers must go on the outer container otherwise you cannot start a selection when the inner container is scrolled
and you click into the overflow area, but the rectangle that is created must go in the inner container itself because it
should overflow inside the scrollable area not outside on the whole page.